refactor: move protocols path to data path
