+++ src/data/usecases/add-account/db-add-account.spec.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { DbAddAccount } from './db-add-account'
-import { Encrypter } from './protocols/encrypter'
+import { Encrypter } from '../../protocols/encrypter'
 
 interface SutTypes {
   sut: DbAddAccount
