+++ src/data/usecases/add-account/db-add-account.ts
@@ -3,7 +3,7 @@ import {
   AddAccount,
   AddAccountModel,
 } from '../../../domain/usecases/add-account'
-import { Encrypter } from './protocols/encrypter'
+import { Encrypter } from '../../protocols/encrypter'
 
 export class DbAddAccount implements AddAccount {
   private readonly encrypter: Encrypter
